# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dcm-anon-vault",
    "description": "Hosted single-tenant DICOM pseudonymization API. Wraps the dcm-anon engine (PS3.15 Basic Confidentiality Profile).",
    "version": "0.3.0"
  },
  "paths": {
...
    "/v1/billing/checkout-session": {
      "post": {
        "summary": "Create Checkout Session",
        "description": "Create a Stripe Checkout session for a Pro/Annual upgrade.",
        "operationId": "create_checkout_session_v1_billing_checkout_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/webhooks": {
      "get": {
        "summary": "List Webhooks",
        "description": "List the calling customer's registered webhook URLs.",
        "operationId": "list_webhooks_v1_webhooks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WebhookResponse"
                  },
                  "type": "array",
                  "title": "Response List Webhooks V1 Webhooks Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Register Webhook",
        "description": "Register a webhook URL for ``anonymize.completed`` notifications.",
        "operationId": "register_webhook_v1_webhooks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookRegistration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/webhooks/deadletter": {
      "get": {
        "summary": "List Deadletter",
        "description": "Admin view of failed webhook deliveries (deadletter queue).",
        "operationId": "list_deadletter_v1_webhooks_deadletter_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeadletterRow"
                  },
                  "title": "Response List Deadletter V1 Webhooks Deadletter Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutRequest": {
        "properties": {
          "success_url": {
            "type": "string",
            "title": "Success Url"
          },
          "cancel_url": {
            "type": "string",
            "title": "Cancel Url"
          },
          "plan": {
            "type": "string",
            "title": "Plan",
            "default": "monthly"
          },
          "customer_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Email"
          }
        },
        "type": "object",
        "required": [
          "success_url",
          "cancel_url"
        ],
        "title": "CheckoutRequest"
      },
      "CheckoutResponse": {
        "properties": {
          "checkout_url": {
            "type": "string",
            "title": "Checkout Url"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "checkout_url",
          "session_id"
        ],
        "title": "CheckoutResponse"
      },
      "DeadletterRow": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "customer_id": {
            "type": "integer",
            "title": "Customer Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "last_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Status"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "url",
          "event_type",
          "last_status",
          "last_error",
          "attempts",
          "created_at"
        ],
        "title": "DeadletterRow"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebhookRegistration": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 512,
            "minLength": 8,
            "title": "Url"
          },
          "secret": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "WebhookRegistration"
      },
      "WebhookResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "active",
          "created_at"
        ],
        "title": "WebhookResponse"
      }
    }
  }
}